hw-5 fixed progress bar, clear form and rest navi yet to fix
