+++ vite-react/src/components/DishCounter/dish-counter.jsx
@@ -1,0 +1,15 @@
+import React, { useState, useCallback } from 'react';
+import { Counter } from "../Counter/counter";
+
+export const DishCounter = () => {
+  const [dishCount, setDishCount] = useState(0);
+
+  const handleDishCount = useCallback((newValue) => {
+    setDishCount(Number(newValue));
+  }, []);
+  return (
+    <div>
+      <Counter value={dishCount} onChange={handleDishCount} />
+    </div>
+  );
+};
