+++ vite-react/src/components/ProgressBar/progbar.module.css
@@ -1,7 +1,17 @@
-.progress-bar {
+.progressBar {
   top: 0;
   left: 0;
+  width: 100%;
   height: 10px;
   position: fixed;
-  background-color: #dc2626;
+  background-color: #fff;
+  background-color: #d1f509;
+  transition: width 0.2s ease;
+  /* border: 1px solid #eee; */
 }
+
+/* .progress-fill {
+  height: 100%;
+  background-color: #d1f509;
+  transition: width 0.2s ease;
+} */
