+++ vite-react/src/components/ProgressBar/scroll-progress.js
@@ -1,8 +1,6 @@
-export const getScrollProgress = () => {
-  return (
-    (window.scrollY /
-      (document.documentElement.scrollHeight -
-        document.documentElement.clientHeight)) *
-    100
-  );
+import React from 'react';
+import './progbar.module.css';
+
+export const ScrollBarProgress = ({ progress }) => {
+  return <div className='progress-fill' style={{ width: `${progress}%` }}></div>
 };
